requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.35.0
peft>=0.7.0
trl>=0.7.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Data Processing
pandas>=2.0.0
datasets>=2.14.0
openpyxl>=3.1.0  # For Excel/CSV if needed

# PDF Processing
pypdf>=3.17.0
pdfminer.six>=20221105

# LLM & RAG
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.1.0
langchain-huggingface>=0.0.1
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4  # Use faiss-gpu if you have CUDA

# API Server
flask>=3.0.0
flask-cors>=4.0.0

# OpenAI API (for synthetic generation)
openai>=1.3.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0

